FIX-0001 SomeViewModelTest (#37)
