+++ core/src/test/java/ru/easycode/words504/BaseTest.kt
